requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg==3.1.18
psycopg-binary==3.1.18
sqlparse==0.5.0
typing_extensions==4.11.0
